+++ CHANGELOG.md
@@ -25,6 +25,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [3.4.1] 2022-01-14
+
+### Added
+
+- Added props (classNameTitle and classNameContent) to allow changing the styles of title and content from the `ModuleCard` component.
+- Added filter props (filter variant and classNameFilter) on `ModuleCard` to display a linear gradient overlay on the top of the image.
+- Added `onPressButton` callback to `ModuleCard` button.
+
+### Fixed
+
+- Fixed problem with `ModuleCard` not having a button callback.
+
 ## [3.4.0] 2022-01-10
 
 ### Added